Editing a Cabin

+++ src/services/apiCabins.js
@@ -1,7 +1,10 @@
 import supabase, { supabaseUrl } from './supabase'
 
 export async function getCabins() {
-  const { data, error } = await supabase.from('cabins').select('*')
+  const { data, error } = await supabase
+    .from('cabins')
+    .select('*')
+    .order('created_at', { ascending: true })
 
   if (error) {
     console.error(error)
@@ -11,35 +14,47 @@ export async function getCabins() {
   return data
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl)
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     '/',
     ''
   )
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`
 
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select()
+  // 1. Create/edit cabin
+  let query = supabase.from('cabins')
+  // A) CREATE
+  if (!id) {
+    query = query.insert([{ ...newCabin, image: imagePath }])
+  }
+  // B) EDIT
+  if (id) {
+    query = query.update({ ...newCabin, image: imagePath }).eq('id', id)
+  }
+  const { data, error } = await query.select()
 
   if (error) {
     console.error(error)
     throw new Error('Cabin could not be created.')
   }
   // 2. Upload image
-  const { error: storageError } = await supabase.storage
-    .from('cabin-images')
-    .upload(imageName, newCabin.image)
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from('cabin-images')
+      .upload(imageName, newCabin.image)
 
-  // 3. Delete the cabin IF there was an error uploading image
-  if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id)
-    console.error(storageError)
-    throw new Error(
-      'Cabins image could not be uploaded and the cabin was not created..'
-    )
+    // 3. Delete the cabin IF there was an error uploading image
+    if (storageError) {
+      await supabase.from('cabins').delete().eq('id', data.id)
+      console.error(storageError)
+      throw new Error(
+        'Cabins image could not be uploaded and the cabin was not created..'
+      )
+    }
   }
 
   return data
